+++ index.js
@@ -1,9 +1,9 @@
-const Device = require('./src/device');
-const { ObjectType, AccessType, DataType, typeToRaw, rawToType, Eds} = require('./src/eds');
-const { EmcyType, EmcyCode, EmcyMessage } = require('./src/protocol/emcy');
-const { LssMode } = require('./src/protocol/lss');
-const { NmtState, NmtCommand } = require('./src/protocol/nmt');
-const { AbortCode, SdoError } = require('./src/protocol/sdo');
+const Device = require('./source/device');
+const { ObjectType, AccessType, DataType, typeToRaw, rawToType, Eds} = require('./source/eds');
+const { EmcyType, EmcyCode, EmcyMessage } = require('./source/protocol/emcy');
+const { LssMode } = require('./source/protocol/lss');
+const { NmtState, NmtCommand } = require('./source/protocol/nmt');
+const { AbortCode, SdoError } = require('./source/protocol/sdo');
 
 module.exports=exports={
     Device,
