+++ source/eds.js
